Move logs/warning to logs/status package
